# src/utils/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "School-ERP APIs",
    "version": "1.0.0",
    "description": "Comprehensive School Management System API with multi-tenant architecture"
  },
...
  "paths":{
...
"/admin/modules": {
    "get": {
      "tags": ["SuperAdmin"],
      "summary": "Get all modules",
      "description": "Fetch all modules available for SuperAdmin",
      "operationId": "getAllModules",
      "responses": {
        "200": {
          "description": "Modules fetched successfully",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "statusCode": {
                    "type": "integer",
                    "example": 200
                  },
                  "data": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Module"
                    }
                  },
                  "message": {
                    "type": "string",
                    "example": "Modules fetched successfully"
                  }
                }
              }
            }
          }
        },
        "404": {
          "description": "Modules not found"
        }
      }
    }
},
...
}}